+++ scoreboard/src/main/java/net/apartium/cocoabeans/scoreboard/TeamMode.java
@@ -1,0 +1,5 @@
+package net.apartium.cocoabeans.scoreboard;
+
+public enum TeamMode {
+    CREATE, REMOVE, UPDATE, ADD_PLAYERS, REMOVE_PLAYERS
+}