+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - [commands] Add Contextual Report for "Not Found" Results in `MapBasedParser` as `NoSuchElementInMapResponse`
 - [common] 📺 **Observable** api
 - [commands] Add ContextualMapBasedParser
+- [minecraft] Added minecraft module
+- [spigot] Scoreboard api
 
 ## 0.0.38
 - [commands] Add `WrappedArgumentParser` for better handling of argument parsers